+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusExporter.cs
@@ -38,6 +38,7 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
         Guard.ThrowIfNull(options);
 
         this.ScrapeResponseCacheDurationMilliseconds = options.ScrapeResponseCacheDurationMilliseconds;
+        this.OpenMetricsEnabled = options.OpenMetricsEnabled;
 
         this.CollectionManager = new PrometheusCollectionManager(this);
     }
@@ -62,6 +63,10 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
     internal PrometheusCollectionManager CollectionManager { get; }
 
     internal int ScrapeResponseCacheDurationMilliseconds { get; }
+
+    internal bool OpenMetricsEnabled { get; }
+
+    internal bool? OpenMetricsRequested { get; set; }
 
     /// <inheritdoc/>
     public override ExportResult Export(in Batch<Metric> metrics)