+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/PrometheusHttpListener.cs
@@ -22,6 +22,8 @@ namespace OpenTelemetry.Exporter;
 
 internal sealed class PrometheusHttpListener : IDisposable
 {
+    private const string OpenMetricsMediaType = "application/openmetrics-text";
+
     private readonly PrometheusExporter exporter;
     private readonly HttpListener httpListener = new();
     private readonly object syncObject = new();
@@ -148,7 +150,9 @@ internal sealed class PrometheusHttpListener : IDisposable
     {
         try
         {
-            var collectionResponse = await this.exporter.CollectionManager.EnterCollect().ConfigureAwait(false);
+            var openMetricsRequested = this.exporter.OpenMetricsEnabled && this.AcceptsOpenMetrics(context.Request);
+            var collectionResponse = await this.exporter.CollectionManager.EnterCollect(openMetricsRequested).ConfigureAwait(false);
+
             try
             {
                 context.Response.Headers.Add("Server", string.Empty);
@@ -156,7 +160,9 @@ internal sealed class PrometheusHttpListener : IDisposable
                 {
                     context.Response.StatusCode = 200;
                     context.Response.Headers.Add("Last-Modified", collectionResponse.GeneratedAtUtc.ToString("R"));
-                    context.Response.ContentType = "text/plain; charset=utf-8; version=0.0.4";
+                    context.Response.ContentType = openMetricsRequested
+                        ? "application/openmetrics-text; version=1.0.0; charset=utf-8"
+                        : "text/plain; charset=utf-8; version=0.0.4";
 
                     await context.Response.OutputStream.WriteAsync(collectionResponse.View.Array, 0, collectionResponse.View.Count).ConfigureAwait(false);
                 }
@@ -186,5 +192,29 @@ internal sealed class PrometheusHttpListener : IDisposable
         catch
         {
         }
+    }
+
+    private bool AcceptsOpenMetrics(HttpListenerRequest request)
+    {
+        var requestAccept = request.Headers["Accept"];
+
+        if (string.IsNullOrEmpty(requestAccept))
+        {
+            return false;
+        }
+
+        var acceptTypes = requestAccept.Split(',');
+
+        foreach (var acceptType in acceptTypes)
+        {
+            var acceptSubType = acceptType.Split(';').FirstOrDefault()?.Trim();
+
+            if (acceptSubType == OpenMetricsMediaType)
+            {
+                return true;
+            }
+        }
+
+        return false;
     }
 }