+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/PrometheusHttpListenerOptions.cs
@@ -49,4 +49,9 @@ public class PrometheusHttpListenerOptions
             this.uriPrefixes = value;
         }
     }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether to export OpenMetrics compatible scrape responses. Default value: true.
+    /// </summary>
+    public bool OpenMetricsEnabled { get; set; } = true;
 }