+++ test/OpenTelemetry.Exporter.Prometheus.AspNetCore.Tests/PrometheusExporterMiddlewareTests.cs
@@ -17,6 +17,7 @@
 #if !NETFRAMEWORK
 using System.Diagnostics.Metrics;
 using System.Net;
+using System.Net.Http.Headers;
 using System.Text.RegularExpressions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
@@ -238,6 +239,16 @@ public sealed class PrometheusExporterMiddlewareTests
             registerMeterProvider: false);
     }
 
+    [Fact]
+    public async Task PrometheusExporterMiddlewareIntegration_DisableOpenMetrics()
+    {
+        await RunPrometheusExporterMiddlewareIntegrationTest(
+            "/metrics",
+            app => app.UseOpenTelemetryPrometheusScrapingEndpoint(),
+            configureOptions: o => o.OpenMetricsEnabled = false,
+            useOpenMetrics: false);
+    }
+
     private static async Task RunPrometheusExporterMiddlewareIntegrationTest(
         string path,
         Action<IApplicationBuilder> configure,
@@ -245,7 +256,8 @@ public sealed class PrometheusExporterMiddlewareTests
         Action<HttpResponseMessage> validateResponse = null,
         bool registerMeterProvider = true,
         Action<PrometheusAspNetCoreOptions> configureOptions = null,
-        bool skipMetrics = false)
+        bool skipMetrics = false,
+        bool useOpenMetrics = true)
     {
         using var host = await new HostBuilder()
            .ConfigureWebHost(webBuilder => webBuilder
@@ -284,7 +296,14 @@ public sealed class PrometheusExporterMiddlewareTests
             counter.Add(0.99D, tags);
         }
 
-        using var response = await host.GetTestClient().GetAsync(path);
+        using var client = host.GetTestClient();
+
+        if (useOpenMetrics)
+        {
+            client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/openmetrics-text"));
+        }
+
+        using var response = await client.GetAsync(path);
 
         var endTimestamp = DateTimeOffset.Now.ToUnixTimeMilliseconds();
 
@@ -292,22 +311,31 @@ public sealed class PrometheusExporterMiddlewareTests
         {
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
             Assert.True(response.Content.Headers.Contains("Last-Modified"));
-            Assert.Equal("text/plain; charset=utf-8; version=0.0.4", response.Content.Headers.ContentType.ToString());
+
+            if (useOpenMetrics)
+            {
+                Assert.Equal("application/openmetrics-text; version=1.0.0; charset=utf-8", response.Content.Headers.ContentType.ToString());
+            }
+            else
+            {
+                Assert.Equal("text/plain; charset=utf-8; version=0.0.4", response.Content.Headers.ContentType.ToString());
+            }
 
             string content = await response.Content.ReadAsStringAsync();
 
-            var matches = Regex.Matches(
-                content,
-                ("^"
-                    + "# TYPE counter_double_total counter\n"
-                    + "counter_double_total{key1='value1',key2='value2'} 101.17 (\\d+)\n"
-                    + "\n"
-                    + "# EOF\n"
-                    + "$").Replace('\'', '"'));
+            string expected = useOpenMetrics
+                ? "# TYPE counter_double_total counter\n"
+                  + "counter_double_total{key1='value1',key2='value2'} 101.17 (\\d+\\.\\d{3})\n"
+                  + "# EOF\n"
+                : "# TYPE counter_double_total counter\n"
+                  + "counter_double_total{key1='value1',key2='value2'} 101.17 (\\d+)\n"
+                  + "# EOF\n";
+
+            var matches = Regex.Matches(content, ("^" + expected + "$").Replace('\'', '"'));
 
             Assert.Single(matches);
 
-            var timestamp = long.Parse(matches[0].Groups[1].Value);
+            var timestamp = long.Parse(matches[0].Groups[1].Value.Replace(".", string.Empty));
 
             Assert.True(beginTimestamp <= timestamp && timestamp <= endTimestamp);
         }